move save_corpus to python module
