+++ backend/conftest.py
@@ -6,7 +6,7 @@ from allauth.account.models import EmailAddress
 
 from ianalyzer.elasticsearch import elasticsearch
 from addcorpus.python_corpora.load_corpus import load_corpus_definition
-from addcorpus.save_corpus import load_and_save_all_corpora
+from addcorpus.python_corpora.save_corpus import load_and_save_all_corpora
 from es import es_index as index
 
 # user credentials and logged-in api clients
