+++ backend/corpora/dbnl/tests/test_dbnl_validation.py
@@ -1,5 +1,5 @@
 import os
-from addcorpus.save_corpus import load_and_save_all_corpora
+from addcorpus.python_corpora.save_corpus import load_and_save_all_corpora
 from addcorpus.models import Corpus
 
 here = os.path.abspath(os.path.dirname(__file__))
