+++ backend/corpora/hum_course_descriptions/tests/test_hum_course_descriptions.py
@@ -2,7 +2,7 @@ import os
 import pytest
 
 from corpora.utils_test import corpus_from_api
-from addcorpus.save_corpus import load_and_save_all_corpora
+from addcorpus.python_corpora.save_corpus import load_and_save_all_corpora
 from addcorpus.python_corpora.load_corpus import load_corpus_definition
 
 here = os.path.abspath(os.path.dirname(__file__))
