+++ backend/corpora/parliament/tests/test_validation.py
@@ -1,4 +1,4 @@
-from addcorpus.save_corpus import load_and_save_all_corpora
+from addcorpus.python_corpora.save_corpus import load_and_save_all_corpora
 from addcorpus.models import Corpus
 
 def test_save_parliament_corpora(parliament_corpora_settings, settings, db):
