+++ backend/corpora/rechtspraak/tests/test_rechtspraak.py
@@ -1,5 +1,5 @@
 from operator import itemgetter
-from addcorpus.save_corpus import load_and_save_all_corpora
+from addcorpus.python_corpora.save_corpus import load_and_save_all_corpora
 from addcorpus.models import Corpus
 
 def test_rechtspraak_sources(test_corpus):
