+++ backend/corpora/utils_test.py
@@ -1,4 +1,4 @@
-from addcorpus.save_corpus import load_and_save_all_corpora
+from addcorpus.python_corpora.save_corpus import load_and_save_all_corpora
 
 def corpus_from_api(client):
     '''
